add support for downloading src tarball in rpmbuild.

+++ gitlab/rpmbuild.sh
@@ -12,13 +12,19 @@
 cd $CI_PROJECT_DIR
 export SPEC=$(ls *spec)
 export PKG=$(rpm -q --specfile $SPEC --queryformat "%{name}-%{version}")
-rm -rf /tmp/$PKG
-mkdir /tmp/$PKG
-cp -R * /tmp/$PKG
-cd /tmp
-tar zcvf ~/rpmbuild/SOURCES/$PKG.tar.gz $PKG
-cd $PKG
-cp *spec ~/rpmbuild/SOURCES
+if [ "$USE_SOURCE_TARBALL" == "1" ]; then
+	spectool -g $SPEC
+	mv $PKG.tar.gz ~/rpmbuild/SOURCES
+	cp *spec ~/rpmbuild/SOURCES
+else
+	rm -rf /tmp/$PKG
+	mkdir /tmp/$PKG
+	cp -R * /tmp/$PKG
+	cd /tmp
+	tar zcvf ~/rpmbuild/SOURCES/$PKG.tar.gz $PKG
+	cd $PKG
+	cp *spec ~/rpmbuild/SOURCES
+fi
 yum-builddep -y $SPEC
 rpmbuild -bs $SPEC
 rpmbuild -bb $SPEC
